Add cone and cylinder

+++ include/camera.h
@@ -10,19 +10,19 @@ class camera {
 		vector3 lookat;
 		vector3 eye;
 		vector3 up;
-		float fov;
-		float near;
-		float far;
+		double fov;
+		double near;
+		double far;
 	public:
 		camera();
-		camera(vector3 Lookat, vector3 Eye, vector3 Up, float Fov, float Near, float Far);
+		camera(vector3 Lookat, vector3 Eye, vector3 Up, double Fov, double Near, double Far);
 		vector3 get_lookat() const;
 		vector3 get_eye() const;
 		vector3 get_up() const;
-		float get_fov() const;
-		float get_near() const;
-		float get_far() const;
-		ray generate_ray(float x, float y, int w, int h);
+		double get_fov() const;
+		double get_near() const;
+		double get_far() const;
+		ray generate_ray(double x, double y, int w, int h);
 		void print(ostream &stream) const;
 };
