+++ include/ray.h
@@ -8,13 +8,13 @@ class ray {
 	public:
 		vector3 origin;
 		vector3 direction;
-		float min_t;
-		float max_t;
+		double min_t;
+		double max_t;
 		ray();
-		ray(vector3 Origin, vector3 Direction, float Min_t, float Max_t);
+		ray(vector3 Origin, vector3 Direction, double Min_t, double Max_t);
 		vector3 get_origin() const;
 		vector3 get_direction() const;
-		float get_max_t() const;
-		float get_min_t() const;
-		vector3 pointAt(float t) const;
+		double get_max_t() const;
+		double get_min_t() const;
+		vector3 pointAt(double t) const;
 };
