+++ src/image.cpp
@@ -3,18 +3,18 @@
 image :: image(){
 }
 
-image :: image(float Width, float Height, color Bgcolor){
+image :: image(double Width, double Height, color Bgcolor){
 	width = Width;
 	height = Height;
 	bgcolor = Bgcolor;
 	data = new char[int(width)*int(height)*3]; 
 }
 
-float image :: get_width() const{
+double image :: get_width() const{
 	return width;
 }
 
-float image :: get_height() const{
+double image :: get_height() const{
 	return height;
 }
 
@@ -22,15 +22,15 @@ color image :: get_bgcolor() const{
 	return bgcolor;
 }
 
-color image :: get_pixel(float x, float y) const{
+color image :: get_pixel(double x, double y) const{
 	int pos=(y*width+x)*3;
-	float r=float(data[pos])/255.0f;
-	float g=float(data[pos+1])/255.0f;
-	float b=float(data[pos+2])/255.0f;
+	double r=double(data[pos])/255.0f;
+	double g=double(data[pos+1])/255.0f;
+	double b=double(data[pos+2])/255.0f;
 	return color(r,g,b);
 }
 		
-void image :: set_pixel(float x, float y, color clr){
+void image :: set_pixel(double x, double y, color clr){
 	int pos=(y*width+x)*3;
 	data[pos] = char(pow(clr.r(),1/2.2)*255+.5);
 	data[pos+1] = char(pow(clr.g(),1/2.2)*255+.5);
