+++ src/ray.cpp
@@ -4,7 +4,7 @@ ray :: ray(){
 
 }
 	
-ray :: ray(vector3 Origin, vector3 Direction, float Min_t, float Max_t){
+ray :: ray(vector3 Origin, vector3 Direction, double Min_t, double Max_t){
 	origin = Origin;
 	direction = Direction;
 	min_t = Min_t;
@@ -19,14 +19,14 @@ vector3 ray :: get_direction() const{
 	return direction;
 }
 
-float ray :: get_max_t() const{
+double ray :: get_max_t() const{
 	return max_t;
 }
 	
-float ray :: get_min_t() const{
+double ray :: get_min_t() const{
 	return min_t;
 }
 	
-vector3 ray :: pointAt(float t) const{
+vector3 ray :: pointAt(double t) const{
 	return origin + direction*t;
 }
